refactor(orchestrator): merge EnhancedMultiAgentOrchestrator into ConsolidatedOrchestrator (#5040)
MERGE PLAN SUMMARY:
- EnhancedMultiAgentOrchestrator had 23 methods; 22 unique, 1 name-collision
- Unique methods ported into ConsolidatedOrchestrator:
  create_workflow_plan, execute_workflow, _execute_single_agent_task,
  get_agent_recommendations, get_performance_report, _calculate_capability_coverage,
  _build_planning_prompt, _parse_planning_response, _evaluate_workflow_criteria,
  _handle_workflow_execution_success, _handle_workflow_execution_failure,
  _handle_task_timeout, _handle_task_exception, _coordinate_collaboration,
  _broadcast_to_agents, _update_workflow_performance_metrics, _publish_workflow_event,
  _get_agent_instance, _get_ma_strategy_handler, _ensure_ma_redis
- Duplicate: _update_agent_performance — both classes had it but with different
  data structures (AgentProfile vs AgentPerformance). Ported version renamed to
  _update_task_agent_performance to preserve both behaviours unambiguously.
- create_and_execute_workflow() convenience fn moved to orchestrator.py
- enhanced_orchestration/__init__.py trimmed to re-exports only (<60 lines)
- api/orchestration.py updated to use get_orchestrator_sync() instead of the
  removed enhanced_orchestrator module-level singleton
- ResourceFactory.get_enhanced_orchestrator → get_orchestrator (with compat alias)
  cache key "enhanced_orchestrator" → "orchestrator" in get_all_cached_resources
- All 34 orchestration tests pass
